# api/openapi/openapi.gen.yaml
...
openapi: 3.1.0
info:
  title: Illarin API
  version: 0.1.0
  description: >-
    State-changing requests that use the session cookie also carry Illarin's
    browser-mutation proof and exact browser origin.
servers:
  - url: /api
paths:
...
  /v1/account/profile:
    put:
      operationId: savePublicProfile
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveProfileRequest'
      responses:
        '200':
          description: The public profile as visitors now see it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: A profile field is not valid
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
  /v1/account/profile/avatar:
    put:
      operationId: setProfileAvatar
      security:
        - sessionCookie: []
          browserMutation: []
      requestBody:
        required: true
        description: Send the image as the only form data part.
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: The public profile carrying its new avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: The image is not valid
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '413':
          description: The image exceeds the upload limit
        '503':
          description: The storage reserve cannot accept the image
    delete:
      operationId: removeProfileAvatar
      security:
        - sessionCookie: []
          browserMutation: []
      responses:
        '200':
          description: The public profile without an avatar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: No account is signed in
        '403':
          description: The signed-in account has not verified its email
        '404':
          description: The profile has no avatar
...
  /download/{id}:
    get:
      operationId: downloadSource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The current uploaded source file, byte for byte
          headers:
            Content-Disposition:
              required: true
              schema:
                type: string
                enum:
                  - attachment
                  - inline
            X-Content-Type-Options:
              required: true
              schema:
                type: string
                enum:
                  - nosniff
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '404':
          description: No such asset
  /download/{id}/{target}:
    get:
      operationId: downloadExport
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: target
          in: path
          required: true
          description: >-
            One of the formats the asset is offered in. A format outside that
            list was never a choice and is refused.
          schema:
            type: string
      responses:
        '200':
          description: >-
            The asset written in the requested format. It is produced on request
            and never cached.
          headers:
            Content-Disposition:
              required: true
              schema:
                type: string
                enum:
                  - attachment
            X-Content-Type-Options:
              required: true
              schema:
                type: string
                enum:
                  - nosniff
            X-Illarin-Export-Target:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: string
                format: binary
            application/zip:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: No such asset
...
  /media/{media_id}/{variant}/{derivative_version}:
    get:
      operationId: getMediaVariant
      parameters:
        - name: media_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: variant
          in: path
          required: true
          schema:
            type: string
            enum:
              - grid
              - grid_blurred
              - detail
              - detail_blurred
              - thumb
              - thumb_blurred
              - og
              - og_blurred
        - name: derivative_version
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: expires
          in: query
          description: >-
            When the signature runs out. A draft's images are served against a
            short-lived signature at the same address they keep once published.
          schema:
            type: string
        - name: signature
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A named media variant
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
            Content-Disposition:
              required: true
              schema:
                type: string
                enum:
                  - inline
            X-Content-Type-Options:
              required: true
              schema:
                type: string
                enum:
                  - nosniff
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: No such media variant
        '503':
          description: The image is temporarily unavailable because storage is low
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: illarin_session
    browserMutation:
      type: apiKey
      in: header
      name: X-Illarin-Request
      description: >-
        Required with the session cookie on state-changing browser requests. The
        value must be 1 and the Origin must be Illarin's browser origin.
    instanceToken:
      type: http
      scheme: bearer
      description: A short-lived access token issued to one linked instance.
...
  schemas:
...
    Profile:
      type: object
      required:
        - id
        - handle
        - displayName
        - biography
        - contactEmail
        - links
        - positions
        - titles
        - badges
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        displayName:
          type: string
        biography:
          type: string
        contactEmail:
          type: string
        avatar:
          anyOf:
            - $ref: '#/components/schemas/ProfileAvatar'
            - type: 'null'
        links:
          type: array
          items:
            $ref: '#/components/schemas/ProfileLink'
        positions:
          type: array
          items:
            $ref: '#/components/schemas/ProfileDistinction'
        titles:
          type: array
          items:
            $ref: '#/components/schemas/ProfileDistinction'
        badges:
          type: array
          items:
            $ref: '#/components/schemas/ProfileDistinction'
...
    DistinctionMark:
      type: object
      required:
        - url
        - width
        - height
      properties:
        url:
          type: string
        width:
          type: integer
        height:
          type: integer
    ProfileDistinction:
      type: object
      required:
        - id
        - name
        - explanation
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        explanation:
          type: string
        mark:
          anyOf:
            - $ref: '#/components/schemas/DistinctionMark'
            - type: 'null'
...
    ProfileLink:
      type: object
      required:
        - label
        - address
      properties:
        label:
          type: string
        address:
          type: string
    SaveProfileRequest:
      type: object
      additionalProperties: false
      required:
        - displayName
        - biography
        - contactEmail
        - links
      properties:
        displayName:
          type: string
          maxLength: 48
        biography:
          type: string
          maxLength: 400
        contactEmail:
          type: string
          maxLength: 254
        links:
          type: array
          maxItems: 6
          items:
            $ref: '#/components/schemas/ProfileLink'
    ProfileAvatar:
      type: object
      required:
        - url
        - width
        - height
      properties:
        url:
          type: string
        width:
          type: integer
        height:
          type: integer
...